requirements.txt
aniso8601==6.0.0
astroid==2.2.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.3
google-api-core==1.14.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-language==1.2.0
googleapis-common-protos==1.6.0
graphene==2.1.7
graphql-core==2.2.1
graphql-relay==2.0.0
grpcio==1.22.0
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.16.4
pandas==0.24.2
praw==6.3.1
prawcore==1.0.1
promise==2.2.1
protobuf==3.9.0
psaw==0.0.7
pyasn1==0.4.5
pyasn1-modules==0.2.5
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
update-checker==0.16
uritemplate==3.0.0
urllib3==1.25.3
websocket-client==0.56.0
whitenoise==4.1.3
wrapt==1.11.2
